Fixed the propertyView width problem.

+++ swing/application/src/main/java/org/jspresso/framework/view/swing/DefaultSwingViewFactory.java
@@ -450,7 +450,7 @@ public class DefaultSwingViewFactory extends
         propertyLabel.setText(" ");
       }
 
-      int propertyWidth = propertyViewDescriptor.getWidth();
+      int propertyWidth = propertyViewDescriptor.getWidth().intValue();
       if (propertyWidth > viewDescriptor.getColumnCount()) {
         propertyWidth = viewDescriptor.getColumnCount();
       }
