+++ ulc/application/src/main/java/org/jspresso/framework/view/ulc/DefaultUlcViewFactory.java
@@ -466,7 +466,7 @@ public class DefaultUlcViewFactory extends
         propertyLabel.setText(" ");
       }
 
-      int propertyWidth = propertyViewDescriptor.getWidth();
+      int propertyWidth = propertyViewDescriptor.getWidth().intValue();
       if (propertyWidth > viewDescriptor.getColumnCount()) {
         propertyWidth = viewDescriptor.getColumnCount();
       }
